requirements.txt
gradio
torchvision
requests
py3Dmol
biopython
pandas
